% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pdbChecks.R
\name{check_lig_name}
\alias{check_lig_name}
\title{Check if target ligand is in the PDB file}
\usage{
check_lig_name(pdb_input, target_ligand)
}
\arguments{
\item{pdb_input}{Pdb path}

\item{target_ligand}{target ligand name}
}
\description{
Check if target ligand is in the PDB file
}
