% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pdbChecks.R
\name{check_residues_name}
\alias{check_residues_name}
\title{Check if residues name are in the PDB file}
\usage{
check_residues_name(pdb_input, target_residues)
}
\arguments{
\item{pdb_input}{Pdb path}

\item{target_residues}{list of residues name}
}
\description{
Check if residues name are in the PDB file
}
\examples{

}
