% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_checkResults.R
\name{check_results}
\alias{check_results}
\title{Function that check and ouput results in the KVFinder interface}
\usage{
check_results(input, output, run_id, is_pg2, url_address, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{run_id}{current run id}

\item{is_pg2}{logical TRUE/FALSE. If TRUE, we calling to create result in page 2 (get latest results page).}

\item{url_address}{url address to conexion}

\item{session}{}
}
\description{
Function that check and ouput results in the KVFinder interface
}
