% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ui_radiobutton.R
\name{create_radio_mode}
\alias{create_radio_mode}
\title{Create radio buttons of the run mode in main Run KVFinder page}
\usage{
create_radio_mode()
}
\description{
Create radio buttons of the run mode in main Run KVFinder page
}
