% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pdbProcess_.R
\name{deal_sele_nonstand}
\alias{deal_sele_nonstand}
\title{Process PDB file according to the selected non-standard residues}
\usage{
deal_sele_nonstand(pdb_input, nonstand_list, include_list)
}
\arguments{
\item{pdb_input}{Pdb path}

\item{nonstand_list}{List of non-standard residues}

\item{include_list}{List of non-standard residues that must be maintained in the PDB}
}
\description{
Process PDB file according to the selected non-standard residues
}
\examples{

}
