% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kvParams.R
\name{kv_params}
\alias{kv_params}
\title{Function that creates radio buttons of the KVFinder parameters}
\usage{
kv_params(id, input_type)
}
\description{
Function that creates radio buttons of the KVFinder parameters
}
