% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kvTutorial.R
\name{kv_tutorial}
\alias{kv_tutorial}
\title{Function that creates the tutorial page of the KVFinder server}
\usage{
kv_tutorial()
}
\description{
Function that creates the tutorial page of the KVFinder server
}
