% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_pdbProcess.R
\name{pdb_process}
\alias{pdb_process}
\title{Steps to process uploaded or fetched PDB}
\usage{
pdb_process(input, output, get_nonstand, mode)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{get_nonstand}{get_nonstand}

\item{mode}{upload or fetch}
}
\description{
Steps to process uploaded or fetched PDB
}
