% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_fetch.R
\name{process_fetch}
\alias{process_fetch}
\title{Steps to process  PDB from fetch in default, customized or box mode}
\usage{
process_fetch(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\description{
Steps to process  PDB from fetch in default, customized or box mode
}
