% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_upload.R
\name{process_upload}
\alias{process_upload}
\title{Function that creates the boxes and buttons of the run mode in upload mode}
\usage{
process_upload(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{}
}
\description{
Function that creates the boxes and buttons of the run mode in upload mode
}
