% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_uploadLigmode.R
\name{process_upload_ligmode}
\alias{process_upload_ligmode}
\title{Steps to process uploaded PDB in ligand mode}
\usage{
process_upload_ligmode(input, output)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}
}
\description{
Steps to process uploaded PDB in ligand mode
}
