% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_snapshot.R
\name{take_snapshot}
\alias{take_snapshot}
\title{Take a snapshot}
\usage{
take_snapshot(input, output, is_pg2)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{is_pg2}{logical TRUE/FALSE. If TRUE, we calling to create result in page 2 (get latest results page).}
}
\description{
Take a snapshot
}
