% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_changeSTRcolor_scheme.R
\name{change_str_color_scheme}
\alias{change_str_color_scheme}
\title{Change color scheme of biomolecular structure}
\usage{
change_str_color_scheme(input, output, protein_col_scheme_list, is_pg2)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{protein_col_scheme_list}{list of previous selected colors}

\item{is_pg2}{logical TRUE/FALSE. If TRUE, we calling to create result in page 2 (get latest results page).}
}
\description{
Change color scheme of biomolecular structure
}
