% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_server_submitJob.R
\name{submit_job}
\alias{submit_job}
\title{Submit job}
\usage{
submit_job(input, output, pdb_name_click_load)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{pdb_name_click_load}{pdb_name_click_load}
}
\description{
Submit job
}
